LayoutTests/editing/selection/5497643-expected.txt

ALERT: SUCCESS: Selection is set to position 2 of BODY.
This tests to make sure that a selection inside a textarea is updated when the textarea is removed from the document.
